AIBookshelf-Bot/handlers/start.py
from telegram import Update
from telegram.ext import ContextTypes

async def start_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Send a message when the command /start is issued."""
    await update.message.reply_text('Welcome to AI Bookshelf! Send me a PDF file and I\'ll help you analyze it.')

async def help_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Send a message when the command /help is issued."""
    await update.message.reply_text('Send me a PDF file and I\'ll help you analyze it.')
