AIBookshelf-Bot/requirements.txt
python-telegram-bot==20.7
PyMySQL==1.1.0
python-dotenv==1.0.0
